% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Image.R
\docType{methods}
\name{addMask,Image-method}
\alias{addMask,Image-method}
\title{Add Mask ROIs to an Image. Expects a binary
mask array binmask covering the whole image.}
\usage{
\S4method{addMask}{Image}(image, c = 0, z = 0, t = 0, binmask)
}
\arguments{
\item{image}{The image}

\item{c}{The channel (default:0 (all))}

\item{z}{The Z plane (default: 0 (all))}

\item{t}{The timepoint (default: 0 (all))}

\item{binmask}{The mask as binary array}
}
\description{
Add Mask ROIs to an Image. Expects a binary
mask array binmask covering the whole image.
}
