% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Image.R
\name{addMask}
\alias{addMask}
\title{Add a Mask ROI to an Image}
\usage{
addMask(image, c = 0, z = 0, t = 0, binmask)
}
\arguments{
\item{image}{The image}

\item{c}{The channel (default:0 (all))}

\item{z}{The Z plane (default: 0 (all))}

\item{t}{The timepoint (default: 0 (all))}

\item{binmask}{The mask as binary array}
}
\description{
Add a Mask ROI to an Image
}
\examples{
\dontrun{
addMask(image, binmask = maskdata)
}
}
