% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Image.R
\name{addROIs}
\alias{addROIs}
\title{Add ROIs to an Image}
\usage{
addROIs(image, z = 1, t = 1, df)
}
\arguments{
\item{image}{The image}

\item{z}{The Z plane (default: 1)}

\item{t}{The timepoint (default: 1)}

\item{df}{The ROI coordinates (x, y)}
}
\description{
Add ROIs to an Image
}
