% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMERO.R
\docType{methods}
\name{attachDataframe,OMERO-method}
\alias{attachDataframe,OMERO-method}
\title{Attaches a dataframe to an OME object}
\usage{
\S4method{attachDataframe}{OMERO}(omero, df, name = "R Dataframe",
  ns = NA)
}
\arguments{
\item{omero}{The OME object}

\item{df}{The dataframe}

\item{name}{An optional name}

\item{ns}{An optional namespace for the file annotation}
}
\value{
The OME object
}
\description{
Attaches a dataframe to an OME object
}
