% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMERO.R
\docType{methods}
\name{attachFile,OMERO-method}
\alias{attachFile,OMERO-method}
\title{Attach a file to an OME object}
\usage{
\S4method{attachFile}{OMERO}(omero, file)
}
\arguments{
\item{omero}{The OME object}

\item{file}{The path to the file to attach}
}
\value{
The file annotation
}
\description{
Attach a file to an OME object
}
