% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMERO.R
\name{attachFile}
\alias{attachFile}
\title{Attach a file to an OME object}
\usage{
attachFile(omero, file)
}
\arguments{
\item{omero}{The OME object}

\item{file}{The path to the file to attach}
}
\value{
The file annotation
}
\description{
Attach a file to an OME object
}
\examples{
\dontrun{
file_annotation <- attachFile(project, file)
}
}
