% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMERO.R
\docType{methods}
\name{availableDataframes,OMERO-method}
\alias{availableDataframes,OMERO-method}
\title{Get the dataframes (name/id/annotationID) attached to an OME object}
\usage{
\S4method{availableDataframes}{OMERO}(omero)
}
\arguments{
\item{omero}{The OME object}
}
\value{
The names/ids of the attached dataframes
}
\description{
Get the dataframes (name/id/annotationID) attached to an OME object
}
