% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMERO.R
\name{availableDataframes}
\alias{availableDataframes}
\title{Get the dataframes (name/id/annotationID) attached to an OME object}
\usage{
availableDataframes(omero)
}
\arguments{
\item{omero}{The OME object}
}
\value{
The names/ids of the attached dataframes
}
\description{
Get the dataframes (name/id/annotationID) attached to an OME object
}
\examples{
\dontrun{
availableDataframes(project)
}
}
