% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMERO.R
\name{cast}
\alias{cast}
\title{Casts a general OMERO object into its proper
type, e. g. Plate (if possible)}
\usage{
cast(omero)
}
\arguments{
\item{omero}{The OME object}
}
\value{
The OMERO object casted to the proper type
}
\description{
Casts a general OMERO object into its proper
type, e. g. Plate (if possible)
}
\examples{
\dontrun{
plate <- cast(object)
}
}
