% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMEROServer.R
\docType{methods}
\name{connect,OMEROServer-method}
\alias{connect,OMEROServer-method}
\title{Connect to an OMERO server}
\usage{
\S4method{connect}{OMEROServer}(server)
}
\arguments{
\item{server}{The server}
}
\value{
The server in "connected" state (if successful)
}
\description{
Connect to an OMERO server
}
