% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMEROServer.R
\name{connect}
\alias{connect}
\title{Connect to an OMERO server}
\usage{
connect(server)
}
\arguments{
\item{server}{The server}
}
\value{
The server in "connected" state (if successful)
}
\description{
Connect to an OMERO server
}
