% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset.R
\docType{class}
\name{Dataset-class}
\alias{Dataset-class}
\alias{Dataset}
\title{Dataset}
\description{
Dataset
}
\section{Slots}{

\describe{
\item{\code{server}}{Reference to the server}

\item{\code{dataobject}}{The Java data object}
}}

