% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMERO.R
\docType{methods}
\name{delete,OMERO-method}
\alias{delete,OMERO-method}
\title{Deletes an OME object}
\usage{
\S4method{delete}{OMERO}(omero)
}
\arguments{
\item{omero}{The OME object}
}
\description{
Deletes an OME object
}
