% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMERO.R
\docType{methods}
\name{deleteFile,OMERO-method}
\alias{deleteFile,OMERO-method}
\title{Delete a file attachend to an OME object}
\usage{
\S4method{deleteFile}{OMERO}(omero, id)
}
\arguments{
\item{omero}{The OME object}

\item{id}{The file annotation(!) id}
}
\description{
Delete a file attachend to an OME object
}
