% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMERO.R
\name{deleteFile}
\alias{deleteFile}
\title{Delete a file attached to an OME object}
\usage{
deleteFile(omero, id)
}
\arguments{
\item{omero}{The OME object}

\item{id}{The file annotation(!) id}
}
\description{
Delete a file attached to an OME object
}
\examples{
\dontrun{
deleteFile(project, file_annotation_id)
}
}
