% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMERO.R
\name{describeDataframe}
\alias{describeDataframe}
\title{Describe a dataframe attached to an OME object}
\usage{
describeDataframe(omero, id)
}
\arguments{
\item{omero}{The OME object}

\item{id}{The id of the dataframe}
}
\value{
NA
}
\description{
Describe a dataframe attached to an OME object
}
\examples{
\dontrun{
describeDataframe(project)
}
}
