% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMEROServer.R
\docType{methods}
\name{disconnect,OMEROServer-method}
\alias{disconnect,OMEROServer-method}
\title{Disconnect from an OMERO server}
\usage{
\S4method{disconnect}{OMEROServer}(server)
}
\arguments{
\item{server}{The server}
}
\value{
The server in "disconnected" state (if successful)
}
\description{
Disconnect from an OMERO server
}
