% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMEROServer.R
\name{disconnect}
\alias{disconnect}
\title{Disconnect from an OMERO server}
\usage{
disconnect(server)
}
\arguments{
\item{server}{The server}
}
\value{
The server in "disconnected" state (if successful)
}
\description{
Disconnect from an OMERO server
}
\examples{
\dontrun{
disconnect(server)
}
}
