% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMERO.R
\docType{methods}
\name{getAnnotations,OMERO-method}
\alias{getAnnotations,OMERO-method}
\title{Get annotations attached to an OME object.
Type, Namespace, Name, Content, ID, FileID (in case of file annotations)}
\usage{
\S4method{getAnnotations}{OMERO}(object, typeFilter, nameFilter)
}
\arguments{
\item{typeFilter}{Optional annotation type filter, e.g. FileAnnotation}

\item{nameFilter}{Optional name filter, e.g. file name of a FileAnnotation}

\item{omero}{The OME object}
}
\value{
The annotations
}
\description{
Get annotations attached to an OME object.
Type, Namespace, Name, Content, ID, FileID (in case of file annotations)
}
