% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMEROServer.R
\docType{methods}
\name{getAnnotations,OMEROServer-method}
\alias{getAnnotations,OMEROServer-method}
\title{Get annotations attached to an OME object.
Type, Namespace, Name, Content, ID, FileID (in case of file annotations)}
\usage{
\S4method{getAnnotations}{OMEROServer}(object, type, id, typeFilter,
  nameFilter)
}
\arguments{
\item{object}{The server}

\item{type}{The object type}

\item{id}{The object id}

\item{typeFilter}{Optional annotation type filter, e.g. FileAnnotation}

\item{nameFilter}{Optional name filter, e.g. file name of a FileAnnotation}
}
\value{
The annotations
}
\description{
Get annotations attached to an OME object.
Type, Namespace, Name, Content, ID, FileID (in case of file annotations)
}
