% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMEROServer.R
\name{getAnnotations}
\alias{getAnnotations}
\title{Get annotations attached to an OME object.
Type, Namespace, Name, Content, ID, FileID (in case of file annotations)}
\usage{
getAnnotations(object, type, id, typeFilter, nameFilter)
}
\arguments{
\item{object}{The OME object}

\item{type}{The object type}

\item{id}{The object id}

\item{typeFilter}{Optional annotation type filter, e.g. FileAnnotation}

\item{nameFilter}{Optional name filter, e.g. file name of a FileAnnotation}
}
\value{
The annotations
}
\description{
Get annotations attached to an OME object.
Type, Namespace, Name, Content, ID, FileID (in case of file annotations)
}
\examples{
\dontrun{
annoations <- getAnnotations(image)
}
}
