% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Image.R
\docType{methods}
\name{getChannelNames,Image-method}
\alias{getChannelNames,Image-method}
\title{Get the channel names of an Image}
\usage{
\S4method{getChannelNames}{Image}(image)
}
\arguments{
\item{image}{The image}
}
\value{
The channel names
}
\description{
Get the channel names of an Image
}
