% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Image.R
\name{getChannelNames}
\alias{getChannelNames}
\title{Get the channel names of an Image}
\usage{
getChannelNames(image)
}
\arguments{
\item{image}{The image}
}
\value{
The channel names
}
\description{
Get the channel names of an Image
}
\examples{
\dontrun{
channelNames <- getChannelNames(image)
}
}
