% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Well.R
\docType{methods}
\name{getColumn,Well-method}
\alias{getColumn,Well-method}
\title{Get the column index of the specific well}
\usage{
\S4method{getColumn}{Well}(well)
}
\arguments{
\item{well}{The well}
}
\value{
The column index
}
\description{
Get the column index of the specific well
}
