% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Well.R
\name{getColumn}
\alias{getColumn}
\title{Get the column index of the specific well}
\usage{
getColumn(well)
}
\arguments{
\item{well}{The well}
}
\value{
The column index
}
\description{
Get the column index of the specific well
}
\examples{
\dontrun{
col <- getColumn(well)
}
}
