% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMEROServer.R
\docType{methods}
\name{getContext,OMEROServer-method}
\alias{getContext,OMEROServer-method}
\title{Get the current SecurityContext}
\usage{
\S4method{getContext}{OMEROServer}(server)
}
\arguments{
\item{server}{The server}
}
\value{
The SecurityContext
}
\description{
Get the current SecurityContext
}
