% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMEROServer.R
\name{getContext}
\alias{getContext}
\title{Get the current SecurityContext}
\usage{
getContext(server)
}
\arguments{
\item{server}{The server}
}
\value{
The SecurityContext
}
\description{
Get the current SecurityContext
}
\examples{
\dontrun{
sec_ctx <- getContext(server)
}
}
