% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMEROServer.R
\docType{methods}
\name{getDatasets,OMEROServer-method}
\alias{getDatasets,OMEROServer-method}
\title{Get all datasets of the logged in user}
\usage{
\S4method{getDatasets}{OMEROServer}(object)
}
\arguments{
\item{object}{The server}
}
\value{
The datasets @seealso \linkS4class{Dataset}
}
\description{
Get all datasets of the logged in user
}
