% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Project.R
\docType{methods}
\name{getDatasets,Project-method}
\alias{getDatasets,Project-method}
\title{Get the datasets of the project}
\usage{
\S4method{getDatasets}{Project}(object)
}
\arguments{
\item{object}{The project}
}
\value{
The datasets (collection of OMERO objects)
}
\description{
Get the datasets of the project
}
