% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMEROServer.R
\name{getDatasets}
\alias{getDatasets}
\title{Get all datasets of the logged in user}
\usage{
getDatasets(object)
}
\arguments{
\item{object}{The server or project}
}
\value{
The datasets @seealso \linkS4class{Dataset}
}
\description{
Get all datasets of the logged in user
}
\examples{
\dontrun{
datasets <- getDatasets(server)
datasets <- getDatasets(project)
}
}
