% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMEROServer.R
\docType{methods}
\name{getGateway,OMEROServer-method}
\alias{getGateway,OMEROServer-method}
\title{Get the reference to the Java Gatway}
\usage{
\S4method{getGateway}{OMEROServer}(server)
}
\arguments{
\item{server}{The server}
}
\value{
The Java Gateway
}
\description{
Get the reference to the Java Gatway
}
