% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMEROServer.R
\name{getGateway}
\alias{getGateway}
\title{Get the reference to the Java Gatway}
\usage{
getGateway(server)
}
\arguments{
\item{server}{The server}
}
\value{
The Java Gateway
}
\description{
Get the reference to the Java Gatway
}
\examples{
\dontrun{
java_gw <- getGateway(server)
}
}
