% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMEROServer.R
\docType{methods}
\name{getGroupContext,OMEROServer-method}
\alias{getGroupContext,OMEROServer-method}
\title{Get the current group context}
\usage{
\S4method{getGroupContext}{OMEROServer}(server)
}
\arguments{
\item{server}{The server}
}
\value{
The name of the group
}
\description{
Get the current group context
}
