% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMEROServer.R
\name{getGroupContext}
\alias{getGroupContext}
\title{Get the current group context}
\usage{
getGroupContext(server)
}
\arguments{
\item{server}{The server}
}
\value{
The name of the group
}
\description{
Get the current group context
}
\examples{
\dontrun{
current_group <- getGroupContext(server)
}
}
