% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plate.R
\docType{methods}
\name{getImages,Plate-method}
\alias{getImages,Plate-method}
\title{Get all images of a plate}
\usage{
\S4method{getImages}{Plate}(omero, fieldIndex)
}
\arguments{
\item{omero}{The plate}

\item{fieldIndex}{Optional field index (only take the 
specific field into account)}
}
\value{
The images as array [[i, j]] where
       i is the well index and j is the field index;
       unless fieldIndex was specified, then a list of
       images is returned. @seealso \linkS4class{Image}
}
\description{
Get all images of a plate
}
\examples{
\dontrun{
images <- getImages(plate)
}
}
