% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Project.R
\docType{methods}
\name{getImages,Project-method}
\alias{getImages,Project-method}
\title{Get all images of a project}
\usage{
\S4method{getImages}{Project}(omero)
}
\arguments{
\item{omero}{The project}
}
\value{
The images @seealso \linkS4class{Image}
}
\description{
Get all images of a project
}
\examples{
\dontrun{
images <- getImages(project)
}
}
