% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Screen.R
\docType{methods}
\name{getImages,Screen-method}
\alias{getImages,Screen-method}
\title{Get all images of the specific screen}
\usage{
\S4method{getImages}{Screen}(omero, fieldIndex = 1)
}
\arguments{
\item{omero}{The screen}

\item{fieldIndex}{The fieldIndex (default = 1)}
}
\value{
The image ids
}
\description{
Get all images of the specific screen
}
