% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Well.R
\docType{methods}
\name{getImages,Well-method}
\alias{getImages,Well-method}
\title{Get the fields of the specific well}
\usage{
\S4method{getImages}{Well}(omero, fieldIndex)
}
\arguments{
\item{omero}{The well}

\item{fieldIndex}{The field index; if specified only the
image of this field is returned}
}
\value{
The field images @seealso \linkS4class{Image}
}
\description{
Get the fields of the specific well
}
\examples{
\dontrun{
images <- getImages(well)
}
}
