% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMERO.R
\name{getImages}
\alias{getImages}
\title{Get all images of a container (Dataset, Project, Screen, Plate or Well)}
\usage{
getImages(omero, fieldIndex)
}
\arguments{
\item{omero}{The OME container object}

\item{fieldIndex}{The field index (only for wells)}
}
\value{
The images @seealso \linkS4class{Image}
}
\description{
Get all images of a container (Dataset, Project, Screen, Plate or Well)
}
\examples{
\dontrun{
images <- getImages(project)
images <- getImages(dataset)
images <- getImages(screen)
}
}
