% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMERO.R
\docType{methods}
\name{getOMEROID,OMERO-method}
\alias{getOMEROID,OMERO-method}
\title{Get the ID of the OME object}
\usage{
\S4method{getOMEROID}{OMERO}(omero)
}
\arguments{
\item{omero}{The OME object}
}
\value{
The OMERO object ID
}
\description{
Get the ID of the OME object
}
