% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMERO.R
\name{getOMEROID}
\alias{getOMEROID}
\title{Get the ID of the OME object}
\usage{
getOMEROID(omero)
}
\arguments{
\item{omero}{The OME object}
}
\value{
The OMERO object ID
}
\description{
Get the ID of the OME object
}
\examples{
\dontrun{
id <- getOMEROID(object)
}
}
