% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMERO.R
\docType{methods}
\name{getOMEROType,OMERO-method}
\alias{getOMEROType,OMERO-method}
\title{Get the type of the OME object}
\usage{
\S4method{getOMEROType}{OMERO}(omero)
}
\arguments{
\item{omero}{The OME object}
}
\value{
The OMERO type
}
\description{
Get the type of the OME object
}
