% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMERO.R
\name{getOMEROType}
\alias{getOMEROType}
\title{Get the type of the OME object}
\usage{
getOMEROType(omero)
}
\arguments{
\item{omero}{The OME object}
}
\value{
The OMERO type
}
\description{
Get the type of the OME object
}
\examples{
\dontrun{
type <- getOMEROType(object)
}
}
