% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Image.R
\name{getPixelValues}
\alias{getPixelValues}
\title{Get the pixel values of an Image}
\usage{
getPixelValues(image, z, t, c)
}
\arguments{
\item{image}{The image}

\item{z}{Z plane index (0 based)}

\item{t}{T plane index (0 based)}

\item{c}{Channel index (0 based)}
}
\value{
The pixel values as two-dimensional array [x][y]
}
\description{
Get the pixel values of an Image
}
