% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMEROServer.R
\docType{methods}
\name{getPlates,OMEROServer-method}
\alias{getPlates,OMEROServer-method}
\title{Get all plates of the logged in user}
\usage{
\S4method{getPlates}{OMEROServer}(object)
}
\arguments{
\item{object}{The server}
}
\value{
The plates @seealso \linkS4class{Plate}
}
\description{
Get all plates of the logged in user
}
