% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Screen.R
\docType{methods}
\name{getPlates,Screen-method}
\alias{getPlates,Screen-method}
\title{Get the plates of the specific screen}
\usage{
\S4method{getPlates}{Screen}(screen)
}
\arguments{
\item{screen}{The screen}
}
\value{
The plates
}
\description{
Get the plates of the specific screen
}
