% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMEROServer.R
\name{getProjects}
\alias{getProjects}
\title{Get all projects of the logged in user}
\usage{
getProjects(server)
}
\arguments{
\item{server}{The server}
}
\value{
The projects @seealso \linkS4class{Project}
}
\description{
Get all projects of the logged in user
}
\examples{
\dontrun{
projects <- getProjects(server)
}
}
