% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Image.R
\docType{methods}
\name{getROIs,Image-method}
\alias{getROIs,Image-method}
\title{Get the ROIs of an Image}
\usage{
\S4method{getROIs}{Image}(image, z = 1, t = 1)
}
\arguments{
\item{image}{The image}

\item{z}{The Z plane (default: 1)}

\item{t}{The timepoint (default: 1)}
}
\value{
The ROIs (x and y coordinates)
}
\description{
Get the ROIs of an Image
}
