% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Well.R
\docType{methods}
\name{getRow,Well-method}
\alias{getRow,Well-method}
\title{Get the row index of the specific well}
\usage{
\S4method{getRow}{Well}(well)
}
\arguments{
\item{well}{The well}
}
\value{
The row index
}
\description{
Get the row index of the specific well
}
