% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Well.R
\name{getRow}
\alias{getRow}
\title{Get the row index of the specific well}
\usage{
getRow(well)
}
\arguments{
\item{well}{The well}
}
\value{
The row index
}
\description{
Get the row index of the specific well
}
\examples{
\dontrun{
row <- getRow(well)
}
}
