% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMEROServer.R
\docType{methods}
\name{getScreens,OMEROServer-method}
\alias{getScreens,OMEROServer-method}
\title{Get all screens of the logged in user}
\usage{
\S4method{getScreens}{OMEROServer}(server)
}
\arguments{
\item{server}{The server}
}
\value{
The screens (collection of OMERO objects)
}
\description{
Get all screens of the logged in user
}
