% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMEROServer.R
\name{getScreens}
\alias{getScreens}
\title{Get all screens of the logged in user}
\usage{
getScreens(server)
}
\arguments{
\item{server}{The server}
}
\value{
The screens @seealso \linkS4class{Screen}
}
\description{
Get all screens of the logged in user
}
\examples{
\dontrun{
screens <- getScreens(server)
}
}
