% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Image.R
\docType{methods}
\name{getThumbnail,Image-method}
\alias{getThumbnail,Image-method}
\title{Get the thumbnail for an Image}
\usage{
\S4method{getThumbnail}{Image}(image)
}
\arguments{
\item{image}{The image}
}
\value{
The thumbnail
}
\description{
Get the thumbnail for an Image
}
