% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Image.R
\name{getThumbnail}
\alias{getThumbnail}
\title{Get the thumbnail for an Image (as JPEG)}
\usage{
getThumbnail(image, width = 96, height = 96)
}
\arguments{
\item{image}{The image}

\item{width}{The width (optional, default = 96)}

\item{height}{The height (optional, default = 96)}
}
\value{
The thumbnail
}
\description{
Get the thumbnail for an Image (as JPEG)
}
\examples{
\dontrun{
jpegThumbnail <- getThumbnail(image)
}
}
