% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMEROServer.R
\docType{methods}
\name{getUser,OMEROServer-method}
\alias{getUser,OMEROServer-method}
\title{Get the current user. This is either the
logged in user or the 'sudo' user if in 
'sudo' mode.}
\usage{
\S4method{getUser}{OMEROServer}(server)
}
\arguments{
\item{server}{The server}
}
\value{
The current user
}
\description{
Get the current user. This is either the
logged in user or the 'sudo' user if in 
'sudo' mode.
}
