% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMEROServer.R
\name{getUser}
\alias{getUser}
\title{Get the current user. This is either the
logged in user or the 'sudo' user if in 
'sudo' mode.}
\usage{
getUser(server)
}
\arguments{
\item{server}{The server}
}
\value{
The current user
}
\description{
Get the current user. This is either the
logged in user or the 'sudo' user if in 
'sudo' mode.
}
\examples{
\dontrun{
user <- getUser(server)
message(paste(user$getUserName(), "is logged in"))
}
}
