% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plate.R
\name{getWells}
\alias{getWells}
\title{Get the wells of the specific plate}
\usage{
getWells(plate)
}
\arguments{
\item{plate}{The plate}
}
\value{
The wells @seealso \linkS4class{Well}
}
\description{
Get the wells of the specific plate
}
\examples{
\dontrun{
wells <- getWells(plate)
}
}
