% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Image.R
\docType{class}
\name{Image-class}
\alias{Image-class}
\alias{Image}
\title{Image}
\description{
Image
}
\section{Slots}{

\describe{
\item{\code{server}}{Reference to the server}

\item{\code{dataobject}}{The Java data object}
}}

\seealso{
\linkS4class{OMERO}
}
