% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Well.R
\docType{methods}
\name{length,Well-method}
\alias{length,Well-method}
\title{Get the number of fields}
\usage{
\S4method{length}{Well}(x)
}
\arguments{
\item{well}{The well}
}
\value{
The number of fields
}
\description{
Get the number of fields
}
