% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMEROServer.R
\docType{methods}
\name{loadCSV,OMEROServer-method}
\alias{loadCSV,OMEROServer-method}
\title{Load a CSV file from the server}
\usage{
\S4method{loadCSV}{OMEROServer}(server, id, header = TRUE, sep = ",",
  quote = "\\"", dec = ".", fill = TRUE, comment.char = "")
}
\arguments{
\item{server}{The server}

\item{id}{The original file ID}

\item{header}{Flag to indicate that the file starts with a header line}

\item{sep}{The separator character}

\item{quote}{The quote character}

\item{dec}{The decimal point character}

\item{fill}{Flag to indicate if blank fields should be added for rows with unequals length}

\item{comment.char}{The comment character}
}
\value{
The dataframe constructed from the CSV file
}
\description{
Load a CSV file from the server
}
