% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMEROServer.R
\docType{methods}
\name{loadObject,OMEROServer-method}
\alias{loadObject,OMEROServer-method}
\title{Load an object from the server}
\usage{
\S4method{loadObject}{OMEROServer}(server, type, id)
}
\arguments{
\item{server}{The server}

\item{type}{The object type}

\item{id}{The object ID}
}
\value{
The OME remote object @seealso \linkS4class{OMERO}
}
\description{
Load an object from the server
}
