% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMEROServer.R
\name{loadObject}
\alias{loadObject}
\title{Load an object from the server}
\usage{
loadObject(server, type, id)
}
\arguments{
\item{server}{The server}

\item{type}{The object type}

\item{id}{The object ID}
}
\value{
The OME remote object @seealso \linkS4class{OMERO}
}
\description{
Load an object from the server
}
\examples{
\dontrun{
obj <- loadObject(server, "ImageData", 123)
image <- cast(obj)
}
}
