% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMERO.R
\docType{class}
\name{OMERO-class}
\alias{OMERO-class}
\alias{OMERO}
\title{OME remote object class
This is basically a wrapper around the Pojo's of the Java Gateway
with a reference to the server.}
\description{
OME remote object class
This is basically a wrapper around the Pojo's of the Java Gateway
with a reference to the server.
}
\section{Slots}{

\describe{
\item{\code{server}}{Reference to the server}

\item{\code{dataobject}}{The Java data object}
}}

