% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMEROServer.R
\docType{class}
\name{OMEROServer-class}
\alias{OMEROServer-class}
\alias{OMEROServer}
\title{OMEROServer class
Provides access to an OMERO server}
\description{
OMEROServer class
Provides access to an OMERO server
}
\section{Slots}{

\describe{
\item{\code{host}}{The host name}

\item{\code{port}}{The port number}

\item{\code{username}}{The username}

\item{\code{password}}{The password}

\item{\code{credentialsFile}}{Text file providing username and password}

\item{\code{gateway}}{Reference to the Gateway}

\item{\code{user}}{The logged in user}

\item{\code{ctx}}{The current SecurityContext}
}}

\examples{
\dontrun{
server <- OMEROServer(username = 'root', password = 'xyz', host = 'localhost')
}
}
