% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plate.R
\docType{class}
\name{Plate-class}
\alias{Plate-class}
\alias{Plate}
\title{Plate}
\description{
Plate
}
\section{Slots}{

\describe{
\item{\code{server}}{Reference to the server}

\item{\code{dataobject}}{The Java data object}
}}

