% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Project.R
\docType{class}
\name{Project-class}
\alias{Project-class}
\alias{Project}
\title{Project}
\description{
Project
}
\section{Slots}{

\describe{
\item{\code{server}}{Reference to the server}

\item{\code{dataobject}}{The Java data object}
}}

