% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Screen.R
\docType{class}
\name{Screen-class}
\alias{Screen-class}
\alias{Screen}
\title{Screen}
\description{
Screen
}
\section{Slots}{

\describe{
\item{\code{server}}{Reference to the server}

\item{\code{dataobject}}{The Java data object}
}}

\seealso{
\linkS4class{OMERO}
}
