% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMEROServer.R
\docType{methods}
\name{setGroup,OMEROServer-method}
\alias{setGroup,OMEROServer-method}
\title{Set a different group context}
\usage{
\S4method{setGroup}{OMEROServer}(server, group)
}
\arguments{
\item{server}{The server}

\item{group}{The name of the group}
}
\value{
The server
}
\description{
Set a different group context
}
