% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMEROServer.R
\name{setGroup}
\alias{setGroup}
\title{Set a different group context}
\usage{
setGroup(server, group)
}
\arguments{
\item{server}{The server}

\item{group}{The name of the group}
}
\value{
The server
}
\description{
Set a different group context
}
