% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMEROServer.R
\name{setGroupContext}
\alias{setGroupContext}
\title{Set a different group context}
\usage{
setGroupContext(server, group)
}
\arguments{
\item{server}{The server}

\item{group}{The name of the group}
}
\value{
The server
}
\description{
Set a different group context
}
\examples{
\dontrun{
server <- setGroupContext(server, 'lab_2')
}
}
