% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMEROServer.R
\name{sudo}
\alias{sudo}
\title{Sudo (act as another user)}
\usage{
sudo(server, username, groupname = NA)
}
\arguments{
\item{server}{The server}

\item{username}{The user to sudo as}

\item{group}{The name of the group to work in}
}
\value{
The server
}
\description{
Sudo (act as another user)
}
\examples{
\dontrun{
server <- sudo(server, 'johnj', 'lab_2')
}
}
