% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Well.R
\docType{class}
\name{Well-class}
\alias{Well-class}
\alias{Well}
\title{Well}
\description{
Well
}
\section{Slots}{

\describe{
\item{\code{server}}{Reference to the server}

\item{\code{dataobject}}{The Java data object}
}}

