% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Image.R
\docType{methods}
\name{addMask,Image-method}
\alias{addMask,Image-method}
\title{Add a Mask ROI to an Image}
\usage{
\S4method{addMask}{Image}(image, z = 1, t = 1, binmask)
}
\arguments{
\item{image}{The image}

\item{z}{The Z plane (default: 1)}

\item{t}{The timepoint (default: 1)}

\item{mask}{The mask as binary array}
}
\description{
Add a Mask ROI to an Image
}
