% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Image.R
\name{addMask}
\alias{addMask}
\title{Add a Mask ROI to an Image}
\usage{
addMask(image, z = 1, t = 1, binmask)
}
\arguments{
\item{image}{The image}

\item{z}{The Z plane (default: 1)}

\item{t}{The timepoint (default: 1)}

\item{binmask}{The mask as binary array}
}
\description{
Add a Mask ROI to an Image
}
\examples{
\dontrun{
addMask(image, binmask = maskdata)
}
}
