% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Image.R
\docType{methods}
\name{addROIs,Image-method}
\alias{addROIs,Image-method}
\title{Add ROIs (Point, Ellipse or Rectangle) to an Image
x, y -> Point
x, y, rx, ry -> Ellipse
x, y, w, h -> Rectangle
theta: Rotation about theta
text: An optional comment}
\usage{
\S4method{addROIs}{Image}(image, z = 1, t = 1, coords)
}
\arguments{
\item{image}{The image}

\item{z}{The Z plane (default: 1)}

\item{t}{The timepoint (default: 1)}

\item{coords}{The coordinates as dataframe (x, y, rx, ry, w, h, theta, text)}
}
\description{
Add ROIs (Point, Ellipse or Rectangle) to an Image
x, y -> Point
x, y, rx, ry -> Ellipse
x, y, w, h -> Rectangle
theta: Rotation about theta
text: An optional comment
}
