% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Image.R
\name{addROIs}
\alias{addROIs}
\title{Add ROIs (Point, Ellipse or Rectangle) to an Image
x, y -> Point
x, y, rx, ry -> Ellipse
x, y, w, h -> Rectangle
theta: Rotation about theta
text: An optional comment}
\usage{
addROIs(image, z = 1, t = 1, coords)
}
\arguments{
\item{image}{The image}

\item{z}{The Z plane (default: 1)}

\item{t}{The timepoint (default: 1)}

\item{coords}{The coordinates as dataframe (x, y, rx, ry, w, h, theta, text)}
}
\description{
Add ROIs (Point, Ellipse or Rectangle) to an Image
x, y -> Point
x, y, rx, ry -> Ellipse
x, y, w, h -> Rectangle
theta: Rotation about theta
text: An optional comment
}
\examples{
\dontrun{
addROIs(image, coords = coordinates)
}
}
