% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMERO.R
\name{attachDataframe}
\alias{attachDataframe}
\title{Attaches a dataframe to an OME object}
\usage{
attachDataframe(omero, df, name = "R Dataframe", ns = NA)
}
\arguments{
\item{omero}{The OME object}

\item{df}{The dataframe}

\item{name}{An optional name}

\item{ns}{An optional namespace for the file annotation}
}
\value{
The OME object
}
\description{
Attaches a dataframe to an OME object
}
\examples{
\dontrun{
attachDataframe(project, df, name="Some data")
}
}
