% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMEROServer.R
\name{connect}
\alias{connect}
\title{Connect to an OMERO server}
\usage{
connect(server, group = NA)
}
\arguments{
\item{server}{The server}

\item{group}{The group context (group name)
(optional, default: user's default group)}
}
\value{
The server in "connected" state (if successful)
}
\description{
Connect to an OMERO server
}
\examples{
\dontrun{
server_connected <- connect(server)
}
}
