% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMERO.R
\docType{methods}
\name{describeDataframe,OMERO-method}
\alias{describeDataframe,OMERO-method}
\title{Describe a dataframe attached to an OME object}
\usage{
\S4method{describeDataframe}{OMERO}(omero, id)
}
\arguments{
\item{omero}{The OME object}

\item{id}{The id of the dataframe}
}
\value{
NA
}
\description{
Describe a dataframe attached to an OME object
}
