% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMERO.R
\docType{methods}
\name{getAnnotations,OMERO-method}
\alias{getAnnotations,OMERO-method}
\title{Get annotations attached to an OME object}
\usage{
\S4method{getAnnotations}{OMERO}(omero, type, name)
}
\arguments{
\item{omero}{The OME object}

\item{type}{The annotation type, e.g. FileAnnotation}

\item{name}{Optional name, e.g. file name of a FileAnnotation}
}
\value{
The annotations
}
\description{
Get annotations attached to an OME object
}
