% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Project.R
\docType{methods}
\name{getDatasets,Project-method}
\alias{getDatasets,Project-method}
\title{Get all datasets of a project}
\usage{
\S4method{getDatasets}{Project}(object)
}
\arguments{
\item{object}{The project}
}
\value{
The datasets @seealso \linkS4class{Dataset}
}
\description{
Get all datasets of a project
}
