% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset.R
\docType{methods}
\name{getImages,Dataset-method}
\alias{getImages,Dataset-method}
\title{Get all images of the dataset}
\usage{
\S4method{getImages}{Dataset}(omero)
}
\arguments{
\item{omero}{The dataset}
}
\value{
The image ids
}
\description{
Get all images of the dataset
}
