% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plate.R
\docType{methods}
\name{getImages,Plate-method}
\alias{getImages,Plate-method}
\title{Get all images of the specific plate}
\usage{
\S4method{getImages}{Plate}(omero)
}
\arguments{
\item{omero}{The plate}
}
\value{
The image ids as matrix [[i, j]] where
       i is the well index and j is the field index
}
\description{
Get all images of the specific plate
}
