% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Screen.R
\docType{methods}
\name{getImages,Screen-method}
\alias{getImages,Screen-method}
\title{Get all images of a screen}
\usage{
\S4method{getImages}{Screen}(omero, fieldIndex = 1)
}
\arguments{
\item{omero}{The screen}

\item{fieldIndex}{The fieldIndex (default = 1)}
}
\value{
A list of image arrays [[i, j]] (one array per plate)
       where i is the well index and j is the field index;
       unless fieldIndex was specified, then a list of
       images is returned. @seealso \linkS4class{Image}
}
\description{
Get all images of a screen
}
\examples{
\dontrun{
images <- getImages(screen)
}
}
