% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Well.R
\docType{methods}
\name{getImages,Well-method}
\alias{getImages,Well-method}
\title{Get the fields of the specific well}
\usage{
\S4method{getImages}{Well}(omero)
}
\arguments{
\item{omero}{The well}
}
\value{
The fields (image ids)
}
\description{
Get the fields of the specific well
}
