% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Image.R
\name{getPixelValues}
\alias{getPixelValues}
\title{Get the pixel values of an Image.
An error will be thrown if invalid z, t or c values 
are specified.}
\usage{
getPixelValues(image, z, t, c)
}
\arguments{
\item{image}{The image}

\item{z}{Z plane index (default: 1)}

\item{t}{T plane index (default: 1)}

\item{c}{Channel index (default: 1)}
}
\value{
The pixel values as two-dimensional array [x][y]
}
\description{
Get the pixel values of an Image.
An error will be thrown if invalid z, t or c values 
are specified.
}
\examples{
\dontrun{
pixel_values <- getPixelValues(image)
}
}
