% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Screen.R
\docType{methods}
\name{getPlates,Screen-method}
\alias{getPlates,Screen-method}
\title{Get the plates of the specific screen}
\usage{
\S4method{getPlates}{Screen}(object)
}
\arguments{
\item{object}{The screen}
}
\value{
The plates @seealso \linkS4class{Plate}
}
\description{
Get the plates of the specific screen
}
