% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMEROServer.R
\docType{methods}
\name{getProjects,OMEROServer-method}
\alias{getProjects,OMEROServer-method}
\title{Get all projects of the logged in user}
\usage{
\S4method{getProjects}{OMEROServer}(server)
}
\arguments{
\item{server}{The server}
}
\value{
The projects @seealso \linkS4class{Project}
}
\description{
Get all projects of the logged in user
}
