% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Image.R
\docType{methods}
\name{getROIs,Image-method}
\alias{getROIs,Image-method}
\title{Get the ROIs (Point, Ellipse or Rectangle) of an Image
x, y -> Point
x, y, rx, ry -> Ellipse
x, y, w, h -> Rectangle
theta: Rotation about theta
text: An optional comment}
\usage{
\S4method{getROIs}{Image}(image, z = 1, t = 1)
}
\arguments{
\item{image}{The image}

\item{z}{The Z plane (default: 1)}

\item{t}{The timepoint (default: 1)}
}
\value{
The coordinates as dataframe (x, y, rx, ry, w, h, theta, text)
}
\description{
Get the ROIs (Point, Ellipse or Rectangle) of an Image
x, y -> Point
x, y, rx, ry -> Ellipse
x, y, w, h -> Rectangle
theta: Rotation about theta
text: An optional comment
}
