% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Image.R
\docType{methods}
\name{getThumbnail,Image-method}
\alias{getThumbnail,Image-method}
\title{Get the thumbnail for an Image}
\usage{
\S4method{getThumbnail}{Image}(image, width = 96, height = 96)
}
\arguments{
\item{image}{The image}

\item{width}{The width (optional, default = 96)}

\item{height}{The height (optional, default = 96)}
}
\value{
The thumbnail
}
\description{
Get the thumbnail for an Image
}
