% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plate.R
\docType{methods}
\name{getWells,Plate-method}
\alias{getWells,Plate-method}
\title{Get the wells of the specific plate}
\usage{
\S4method{getWells}{Plate}(plate)
}
\arguments{
\item{plate}{The plate}
}
\value{
The wells
}
\description{
Get the wells of the specific plate
}
