% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMERO.R
\docType{methods}
\name{loadDataframe,OMERO-method}
\alias{loadDataframe,OMERO-method}
\title{Load a dataframe attached to an OME object}
\usage{
\S4method{loadDataframe}{OMERO}(omero, id, rowFrom, rowTo, columns)
}
\arguments{
\item{omero}{The OME object}

\item{id}{The id of the dataframe}

\item{rowFrom}{Load data from row}

\item{rowTo}{Load data to row}

\item{columns}{Only specified columns}
}
\value{
The dataframe
}
\description{
Load a dataframe attached to an OME object
}
