% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMEROServer.R
\docType{methods}
\name{sudo,OMEROServer-method}
\alias{sudo,OMEROServer-method}
\title{Sudo (act as another user)}
\usage{
\S4method{sudo}{OMEROServer}(server, username, groupname = NA)
}
\arguments{
\item{server}{The server}

\item{username}{The user to sudo as (use NA to disable the sudo context again)}

\item{group}{The name of the group to work in}
}
\value{
The server
}
\description{
Sudo (act as another user)
}
\examples{
\dontrun{
server <- sudo(server, 'johnj', 'lab_2')
}
}
